improve conformance testing utilities
