+++ Firestore/tests/Conformance/FirestoreTest.php
@@ -17,7 +17,6 @@
 
 namespace Google\Cloud\Firestore\Tests\Conformance;
 
-use Google\ApiCore\Serializer;
 use Google\Cloud\Core\Testing\ArrayHasSameValuesToken;
 use Google\Cloud\Core\Testing\GrpcTestTrait;
 use Google\Cloud\Core\Testing\TestHelpers;
@@ -31,8 +30,6 @@ use Google\Cloud\Firestore\FieldValue;
 use Google\Cloud\Firestore\FirestoreClient;
 use Google\Cloud\Firestore\PathTrait;
 use Google\Cloud\Firestore\ValueMapper;
-use Google\Cloud\Tests\Conformance\Firestore\FirestoreTestSuite as Message;
-use Google\Protobuf\Internal\CodedInputStream;
 use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
 use Prophecy\Exception\Call\UnexpectedCallException;
@@ -42,11 +39,11 @@ use Prophecy\Exception\Call\UnexpectedCallException;
  */
 class FirestoreTest extends TestCase
 {
+    use ConformanceTestTrait;
     use GrpcTestTrait;
     use PathTrait;
 
-    private static $cases = [];
-    private static $skipped = [];
+    const SUITE_FILENAME = 'firestore-test-suite.binproto';
 
     private $testTypes = ['get', 'create', 'set', 'update', 'updatePaths', 'delete', 'query'];
     private $client;
@@ -79,7 +76,7 @@ class FirestoreTest extends TestCase
     }
 
     /**
-     * @dataProvider getCases
+     * @dataProvider cases
      * @group firestore-get
      */
     public function testGet($test)
@@ -93,7 +90,7 @@ class FirestoreTest extends TestCase
     }
 
     /**
-     * @dataProvider createCases
+     * @dataProvider cases
      * @group firestore-create
      */
     public function testCreate($test)
@@ -127,7 +124,7 @@ class FirestoreTest extends TestCase
     }
 
     /**
-     * @dataProvider setCases
+     * @dataProvider cases
      * @group firestore-set
      */
     public function testSet($test)
@@ -166,7 +163,7 @@ class FirestoreTest extends TestCase
     }
 
     /**
-     * @dataProvider updateCases
+     * @dataProvider cases
      * @group firestore-update
      */
     public function testUpdate($test)
@@ -209,7 +206,7 @@ class FirestoreTest extends TestCase
     }
 
     /**
-     * @dataProvider updatePathsCases
+     * @dataProvider cases
      * @group firestore-updatepaths
      */
     public function testUpdatePaths($test)
@@ -256,7 +253,7 @@ class FirestoreTest extends TestCase
     }
 
     /**
-     * @dataProvider deleteCases
+     * @dataProvider cases
      * @group firestore-delete
      */
     public function testDelete($test)
@@ -292,7 +289,7 @@ class FirestoreTest extends TestCase
     }
 
     /**
-     * @dataProvider queryCases
+     * @dataProvider cases
      * @group firestore-query
      */
     public function testQuery($test, $desc)
@@ -397,16 +394,6 @@ class FirestoreTest extends TestCase
         }
     }
 
-    /**
-     * Report skipped cases for measurement purposes.
-     *
-     * @dataProvider skippedCases
-     */
-    public function testSkipped($desc)
-    {
-        $this->markTestSkipped($desc);
-    }
-
     private function formatOptions(array $test)
     {
         $options = [];
@@ -462,69 +449,4 @@ class FirestoreTest extends TestCase
 
         return $value;
     }
-
-    public function cases($type)
-    {
-        $cases = array_filter($this->setupCases(), function ($case) use ($type) {
-            return $case['type'] === $type;
-        });
-
-        $res = [];
-        foreach ($cases as $case) {
-            $res[$case['description']] = [$case['test'], $case['description']];
-        }
-
-        return $res;
-    }
-
-    private function setupCases()
-    {
-        if (self::$cases) {
-            return self::$cases;
-        }
-
-        $serializer = new Serializer;
-
-        $str = file_get_contents(__DIR__ . '/proto/firestore-test-suite.binproto');
-        $suite = new Message;
-        $suite->mergeFromString($str);
-
-        $cases = [];
-        foreach ($suite->getTests() as $test) {
-            $case = $serializer->encodeMessage($test);
-            $matches = array_values(array_intersect($this->testTypes, array_keys($case)));
-            if (!$matches) {
-                $keys = array_keys($case);
-                throw new \Exception(sprintf(
-                    'Unknown test type! Keys were `%s`',
-                    implode(', ', $keys)
-                ));
-            }
-
-            $type = $matches[0];
-
-            if (in_array($case['description'], $this->excludes)) {
-                self::$skipped[] = [$case['description']];
-                continue;
-            }
-
-            $cases[] = [
-                'description' => $case['description'],
-                'type' => $type,
-                'test' => $case[$type]
-            ];
-        }
-
-        self::$cases = $cases;
-        return $cases;
-    }
-
-    public function getCases() { return $this->cases('get'); }
-    public function createCases() { return $this->cases('create'); }
-    public function setCases() { return $this->cases('set'); }
-    public function updateCases() { return $this->cases('update'); }
-    public function updatePathsCases() { return $this->cases('updatePaths'); }
-    public function deleteCases() { return $this->cases('delete'); }
-    public function queryCases() { return $this->cases('query'); }
-    public function skippedCases() { return self::$skipped; }
 }
